requirements.txt
streamlit>=1.28.0
pygwalker>=0.4.0
pandas>=2.0.0
tdd-guard-pytest